fix(FR-819): Form problem
Form elements that reference a form by ID can only reference forms that
have already been defined in document order. So when we moved the action
bar from side bar to above table, the action selector was defined before
the form itself, including the submit button and so the form was no
longer submitted.

By defining the form *before* the action bar, we circumvent this issue,
at the price of having to reference every element to the form.