feat(FR-819): Re-align fields in form-row
When readonly field and non-readonly field are combined, it looks
weird, which is why initially a choice was made for flex-end. However,
the root cause is that nothing gets transformed/translated on readonly
fields. So compensate for that and revert back to flex-start, because
that overall works better because of the help-text that is rendered
below the widget.