feat(FR-1412): Implement Static Currency widget
The currency for claim documents will be correct, so display it as a
hint to claims managers and trigger if the invoice is in another
currency.

Since we only allow invoices in the budget's currency, we don't allow
alteration of this value. However, for historical reasons, we should
display the actual value and not the budget currency here. This is why
we need a widget that can display the currency as it was recorded.